+++ test/e2e/components/TestComponent.js
@@ -1,7 +1,7 @@
 const dlog = require ('./../dlog.js');
 
 exports.namedOnExport = function (p1, p2) {
-  dlog.log( { 'namedOnExport' : {p1 : p1, p2 : p2} }, { arguments } )
+  dlog.log( { 'function' : {p1 : p1, p2 : p2} }, { arguments } )
  };
 
 export default (p1, p2) => {