+++ test/e2e/typeChecking.js
@@ -3,7 +3,7 @@ const typCheckingExample = () => {
   dlog.log( { 'typCheckingExample' : {} }, { arguments } )
 
   const wantsAStringFn = function (str) {
-    dlog.log( { 'wantsAStringFn' : {str : str} }, { arguments } )
+  dlog.log( { 'function' : {str : str} }, { arguments } )
 
     return str;
   };